refactor: move shared models to common lib
