+++ libs/common/src/index.ts
@@ -6,3 +6,4 @@ export * from './lib/common.service'
 export * from './lib/color.service'
 export * from './lib/common.module'
 export * from './lib/i18n/i18n.module'
+export * from './lib/model'
