+++ libs/search/src/index.ts
@@ -6,4 +6,3 @@ export * from './lib/search.module'
 export * from './lib/state/actions'
 export * from './lib/state/selectors'
 export * from './lib/state/effects'
-export * from './lib/model'
