+++ libs/search/src/lib/results-list/results-list.container.component.html
@@ -1,4 +1,5 @@
 <ui-results-list
   [loading]="isLoading$ | async"
   [records]="results$ | async"
+  [layout]="layout"
 ></ui-results-list>
