+++ libs/search/src/lib/results-list/results-list.container.component.ts
@@ -1,7 +1,8 @@
-import { Component, OnInit } from '@angular/core'
+import { Component, Input, OnInit } from '@angular/core'
+import { ResultsListLayout } from '@lib/common'
 import { select, Store } from '@ngrx/store'
+import { SearchState } from '../state/reducer'
 import { getSearchResults, getSearchResultsLoading } from '../state/selectors'
-import { SearchState } from '../model'
 import { RequestMoreResults } from '../state/actions'
 
 @Component({
@@ -10,6 +11,8 @@ import { RequestMoreResults } from '../state/actions'
   styleUrls: ['./results-list.container.component.css'],
 })
 export class ResultsListContainerComponent implements OnInit {
+  @Input() layout: ResultsListLayout = ResultsListLayout.BLOCK
+
   results$ = this.store.pipe(select(getSearchResults))
   isLoading$ = this.store.pipe(select(getSearchResultsLoading))
