+++ libs/search/src/lib/state/reducer.ts
@@ -1,7 +1,14 @@
-import { SearchState } from '../model'
+import { RecordSimple, SearchParams } from '@lib/common'
 import * as fromActions from './actions'
 
 export const SEARCH_FEATURE_KEY = 'searchState'
+
+export interface SearchState {
+  params: SearchParams
+  sortBy?: string
+  results: RecordSimple[]
+  loadingMore: boolean
+}
 
 export const initialState: SearchState = {
   params: {},
