+++ libs/search/src/lib/state/selectors.ts
@@ -1,6 +1,5 @@
 import { createFeatureSelector, createSelector } from '@ngrx/store'
-import { SearchState } from '../model'
-import { SEARCH_FEATURE_KEY } from './reducer'
+import { SEARCH_FEATURE_KEY, SearchState } from './reducer'
 
 export const getSearchState = createFeatureSelector<SearchState>(
   SEARCH_FEATURE_KEY
